LayoutTests/plugins/fullscreen-plugins-dont-reload-expected.txt

ALERT: Plugin Loaded!
go fullscreen
There should only be one ALERT. If there were two, the plugin was reloaded during the transition to fullscreen.
